requirements.txt
PyQt6
cryptography
smaz-py3